Update SlimTwigWrapper.php
Allow specifying which template to load, either from the root's view file locations or within the subroot (if in working inside a subroot), if the the file to render is prefixed with a '/' (look in root) or not (look in subroot). If not in a subroot, then the '/' makes no difference.